Loopback Fix

+++ STM32/Src/fpga.c
@@ -304,12 +304,12 @@ void FPGA_fpgadata_getiq(void)
 	FPGA_fpgadata_in_inttmp16 = FPGA_fpgadata_in_tmp16;
 	if(FFTInputBufferInProgress) // A buffer in progress
 	{
-		FFTInput_A[FFT_buff_index + 1] = ((float32_t)FPGA_fpgadata_in_inttmp16/32767);
+		FFTInput_A[FFT_buff_index + 1] = (int16_t)FPGA_fpgadata_in_inttmp16/32767.0f;
 		FPGA_Audio_Buffer_Q[FPGA_Audio_Buffer_Index] = FFTInput_A[FFT_buff_index + 1];
 	}
 	else // B buffer in progress
 	{
-		FFTInput_B[FFT_buff_index + 1] = ((float32_t)FPGA_fpgadata_in_inttmp16/32767);
+		FFTInput_B[FFT_buff_index + 1] = (int16_t)FPGA_fpgadata_in_inttmp16/32767.0f;
 		FPGA_Audio_Buffer_Q[FPGA_Audio_Buffer_Index] = FFTInput_B[FFT_buff_index + 1];
 	}
 	//clock
@@ -347,12 +347,12 @@ void FPGA_fpgadata_getiq(void)
 	FPGA_fpgadata_in_inttmp16 = FPGA_fpgadata_in_tmp16;
 	if(FFTInputBufferInProgress) // A buffer in progress
 	{
-		FFTInput_A[FFT_buff_index] = ((float32_t)FPGA_fpgadata_in_inttmp16/32767);
+		FFTInput_A[FFT_buff_index] = (int16_t)FPGA_fpgadata_in_inttmp16/32767.0f;
 		FPGA_Audio_Buffer_I[FPGA_Audio_Buffer_Index] = FFTInput_A[FFT_buff_index];
 	}
 	else // B buffer in progress
 	{
-		FFTInput_B[FFT_buff_index] = ((float32_t)FPGA_fpgadata_in_inttmp16/32767);
+		FFTInput_B[FFT_buff_index] = (int16_t)FPGA_fpgadata_in_inttmp16/32767.0f;
 		FPGA_Audio_Buffer_I[FPGA_Audio_Buffer_Index] = FFTInput_B[FFT_buff_index];
 	}
 	FPGA_Audio_Buffer_Index++;
